+++ src/test/java/org/stellar/sdk/TransactionTest.java
@@ -506,4 +506,193 @@ public class TransactionTest {
             Network.PUBLIC);
     assertNotEquals(transaction1, transaction4);
   }
+
+  @Test
+  public void testIntegrationPaymentToContractTransactionWithNativeAsset() throws IOException {
+    Asset asset = Asset.createNativeAsset();
+    BigDecimal amount1 = new BigDecimal("100.125");
+    BigDecimal amount2 = new BigDecimal("120.7891");
+
+    KeyPair kp = KeyPair.random();
+    IntegrationUtils.fundAccount(kp.getAccountId());
+    IntegrationUtils.createAssetContract(asset, kp);
+    String destination = IntegrationUtils.getRandomContractId(kp);
+    Server server = new Server(IntegrationUtils.HORIZON_URL);
+    TransactionBuilderAccount sourceAccount = server.loadAccount(kp.getAccountId());
+
+    Transaction transaction1 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildPaymentToContractTransaction(destination, asset, amount1, null);
+    transaction1.sign(kp);
+    server.submitTransaction(transaction1);
+    BigInteger balance1 = IntegrationUtils.getBalanceForContract(destination, asset, kp);
+    assertEquals(1001250000L, balance1.longValue());
+
+    Transaction transaction2 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildPaymentToContractTransaction(destination, asset, amount2, null);
+    transaction2.sign(kp);
+    server.submitTransaction(transaction2);
+    BigInteger balance2 = IntegrationUtils.getBalanceForContract(destination, asset, kp);
+    assertEquals(2209141000L, balance2.longValue());
+
+    Transaction transaction3 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildRestoreAssetBalanceEntryTransaction(destination, asset, null);
+    transaction3.sign(kp);
+    server.submitTransaction(transaction3);
+
+    server.close();
+  }
+
+  @Test
+  public void testIntegrationPaymentToContractTransactionWithAlphanum4Asset() throws IOException {
+    BigDecimal amount1 = new BigDecimal("100.125");
+    BigDecimal amount2 = new BigDecimal("120.7891");
+
+    KeyPair issuerKp = KeyPair.random();
+    KeyPair kp = KeyPair.random();
+    IntegrationUtils.fundAccount(issuerKp.getAccountId());
+    IntegrationUtils.fundAccount(kp.getAccountId());
+    Asset asset = Asset.createNonNativeAsset("CAT", issuerKp.getAccountId());
+    IntegrationUtils.issueAsset("CAT", issuerKp, kp, new BigDecimal("1000"));
+    IntegrationUtils.createAssetContract(asset, kp);
+    String destination = IntegrationUtils.getRandomContractId(kp);
+    Server server = new Server(IntegrationUtils.HORIZON_URL);
+    TransactionBuilderAccount sourceAccount = server.loadAccount(kp.getAccountId());
+
+    Transaction transaction1 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildPaymentToContractTransaction(destination, asset, amount1, null);
+    transaction1.sign(kp);
+    server.submitTransaction(transaction1);
+    BigInteger balance1 = IntegrationUtils.getBalanceForContract(destination, asset, kp);
+    assertEquals(1001250000L, balance1.longValue());
+
+    Transaction transaction2 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildPaymentToContractTransaction(destination, asset, amount2, null);
+    transaction2.sign(kp);
+    server.submitTransaction(transaction2);
+    BigInteger balance2 = IntegrationUtils.getBalanceForContract(destination, asset, kp);
+    assertEquals(2209141000L, balance2.longValue());
+
+    Transaction transaction3 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildRestoreAssetBalanceEntryTransaction(destination, asset, null);
+    transaction3.sign(kp);
+    server.submitTransaction(transaction3);
+
+    server.close();
+  }
+
+  @Test
+  public void testIntegrationPaymentToContractTransactionWithAlphanum12Asset() throws IOException {
+    BigDecimal amount1 = new BigDecimal("100.125");
+    BigDecimal amount2 = new BigDecimal("120.7891");
+
+    KeyPair issuerKp = KeyPair.random();
+    KeyPair kp = KeyPair.random();
+    IntegrationUtils.fundAccount(issuerKp.getAccountId());
+    IntegrationUtils.fundAccount(kp.getAccountId());
+    Asset asset = Asset.createNonNativeAsset("BANANA", issuerKp.getAccountId());
+    IntegrationUtils.issueAsset("BANANA", issuerKp, kp, new BigDecimal("1000"));
+    IntegrationUtils.createAssetContract(asset, kp);
+    String destination = IntegrationUtils.getRandomContractId(kp);
+    Server server = new Server(IntegrationUtils.HORIZON_URL);
+    TransactionBuilderAccount sourceAccount = server.loadAccount(kp.getAccountId());
+
+    Transaction transaction1 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildPaymentToContractTransaction(destination, asset, amount1, null);
+    transaction1.sign(kp);
+    server.submitTransaction(transaction1);
+    BigInteger balance1 = IntegrationUtils.getBalanceForContract(destination, asset, kp);
+    assertEquals(1001250000L, balance1.longValue());
+
+    Transaction transaction2 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildPaymentToContractTransaction(destination, asset, amount2, null);
+    transaction2.sign(kp);
+    server.submitTransaction(transaction2);
+    BigInteger balance2 = IntegrationUtils.getBalanceForContract(destination, asset, kp);
+    assertEquals(2209141000L, balance2.longValue());
+
+    Transaction transaction3 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildRestoreAssetBalanceEntryTransaction(destination, asset, null);
+    transaction3.sign(kp);
+    server.submitTransaction(transaction3);
+
+    server.close();
+  }
+
+  @Test
+  public void testIntegrationPaymentToContractTransactionWithDifferentSource() throws IOException {
+    Asset asset = Asset.createNativeAsset();
+    BigDecimal amount1 = new BigDecimal("100.125");
+    BigDecimal amount2 = new BigDecimal("120.7891");
+
+    KeyPair kp = KeyPair.random();
+    KeyPair opSource = KeyPair.random();
+    IntegrationUtils.fundAccount(kp.getAccountId());
+    IntegrationUtils.fundAccount(opSource.getAccountId());
+    IntegrationUtils.createAssetContract(asset, kp);
+    String destination = IntegrationUtils.getRandomContractId(kp);
+    Server server = new Server(IntegrationUtils.HORIZON_URL);
+    TransactionBuilderAccount sourceAccount = server.loadAccount(kp.getAccountId());
+
+    Transaction transaction1 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildPaymentToContractTransaction(
+                destination, asset, amount1, opSource.getAccountId());
+    transaction1.sign(kp);
+    transaction1.sign(opSource);
+    server.submitTransaction(transaction1);
+    BigInteger balance1 = IntegrationUtils.getBalanceForContract(destination, asset, kp);
+    assertEquals(1001250000L, balance1.longValue());
+
+    Transaction transaction2 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildPaymentToContractTransaction(
+                destination, asset, amount2, opSource.getAccountId());
+    transaction2.sign(kp);
+    transaction2.sign(opSource);
+    server.submitTransaction(transaction2);
+    BigInteger balance2 = IntegrationUtils.getBalanceForContract(destination, asset, kp);
+    assertEquals(2209141000L, balance2.longValue());
+
+    Transaction transaction3 =
+        new TransactionBuilder(sourceAccount, IntegrationUtils.NETWORK)
+            .setBaseFee(100)
+            .setTimeout(300)
+            .buildRestoreAssetBalanceEntryTransaction(destination, asset, opSource.getAccountId());
+    transaction3.sign(kp);
+    transaction3.sign(opSource);
+    server.submitTransaction(transaction3);
+
+    server.close();
+  }
 }